backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
anthropic==0.25.1
django-filter==24.2
requests==2.31.0
gunicorn==21.2.0
